refactor: migrate shared utils into shared/utils with stubs
