+++ packages/commonwealth/client/scripts/helpers/awsHelpers.ts
@@ -1,7 +1,1 @@
-import { isLinkValid } from './link';
-
-const S3_URL_REGEX = /^https:\/\/s3(\.[^.]+)?\.amazonaws\.com(\/.*)?$/;
-
-export function isS3URL(url: string) {
-  return isLinkValid(url) && S3_URL_REGEX.test(url);
-}
+export * from 'shared/utils/awsHelpers';
