+++ packages/commonwealth/client/scripts/helpers/rateLimit.ts
@@ -1,32 +1,1 @@
-import {
-  TierRateLimitErrorMessage,
-  TierRateLimitErrors,
-} from '@hicommonwealth/shared';
-
-export const RATE_LIMIT_MESSAGE =
-  'You are being rate limited. Please wait and try again.';
-
-export interface RateLimitErrorType {
-  data?: { httpStatus?: number; message?: string };
-  status?: number;
-  response?: { status?: number; data?: { message?: string } };
-  message?: string;
-}
-
-export const isRateLimitError = (err: RateLimitErrorType) => {
-  const status = err?.data?.httpStatus || err?.status || err?.response?.status;
-  if (status === 429) return true;
-
-  const msg =
-    err?.data?.message || err?.message || err?.response?.data?.message || '';
-  const lowerMsg = String(msg).toLowerCase();
-  return (
-    lowerMsg.includes('rate limit') || lowerMsg.includes('too many requests')
-  );
-};
-
-export const isTierRateLimitError = (err: Error): boolean => {
-  return Object.values(TierRateLimitErrors).includes(
-    err.message as TierRateLimitErrorMessage,
-  );
-};
+export * from 'shared/utils/rateLimit';
