+++ packages/commonwealth/client/scripts/helpers/tooltipTexts.ts
@@ -1,17 +1,1 @@
-import { ChainBase } from '@hicommonwealth/shared';
-
-type StakeButtonToolTipHelperProps = {
-  isLoggedIn?: boolean;
-  connectBaseChainToBuy?: ChainBase;
-};
-
-export const disabledStakeButtonTooltipText = ({
-  isLoggedIn = true,
-  connectBaseChainToBuy,
-}: StakeButtonToolTipHelperProps) => {
-  if (!isLoggedIn) return 'Login to buy stakes';
-  if (connectBaseChainToBuy) {
-    return `Connect an ${connectBaseChainToBuy} based wallet or join this community to buy stake`;
-  }
-  return '';
-};
+export * from 'shared/utils/tooltipTexts';
