+++ packages/commonwealth/client/scripts/shared/utils/awsHelpers.ts
@@ -1,0 +1,7 @@
+import { isLinkValid } from './link';
+
+const S3_URL_REGEX = /^https:\/\/s3(\.[^.]+)?\.amazonaws\.com(\/.*)?$/;
+
+export function isS3URL(url: string) {
+  return isLinkValid(url) && S3_URL_REGEX.test(url);
+}
