+++ packages/commonwealth/client/scripts/utils/clipboard.ts
@@ -1,13 +1,1 @@
-import { notifySuccess } from 'controllers/app/notifications';
-
-export const saveToClipboard = async (
-  content: string,
-  successNotification = false,
-) => {
-  try {
-    await navigator.clipboard.writeText(content);
-    successNotification && notifySuccess('Copied');
-  } catch (err) {
-    console.log(err);
-  }
-};
+export * from 'shared/utils/clipboard';
